backend/pyproject.toml
[project]
name = "paperrelay-backend"
version = "0.1.0"
description = "Backend services for PaperRelay"
requires-python = ">=3.11,<3.13"
dependencies = [
    "alembic==1.13.0",
    "celery==5.3.6",
    "email-validator==2.1.0",
    "fastapi==0.109.0",
    "httpx==0.26.0",
    "markdown==3.5.1",
    "openai==1.10.0",
    "passlib[bcrypt]==1.7.4",
    "pdfplumber==0.10.3",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pymupdf==1.23.8",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "reportlab==4.0.8",
    "slowapi==0.1.9",
    "sqlalchemy==2.0.25",
    "tenacity==8.2.3",
    "uvicorn[standard]==0.27.0",
]

[dependency-groups]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
